FormFiller App SL/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
entrypoints==0.4
extra-streamlit-components==0.1.71
Faker==24.4.0
favicon==0.7.0
fire==0.6.0
fonttools==4.50.0
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
numpy==1.26.4
openai==1.14.3
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.1
pdf2docx==0.5.8
pillow==10.2.0
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pymdown-extensions==10.7.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.32.2
streamlit-authenticator==0.3.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.3
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
